refactor(program): I/O 層を ProgramFetcher に分離 (#4347)
Program の HTTP 取得・YAML/Redis 永続化を ProgramFetcher へ切り出し、
Program は番組表データの参照・編集に専念。Metrics/ClassLength の
rubocop:disable を解除 (225→134 行)。

- ProgramFetcher: uris/fetch/save/load/yaml_exist?/invalidate_cache/redis
  と検証・キャッシュ系 private を保持
- Program: I/O はメモ化した fetcher へ委譲。YAML_PATH は後方互換で再公開
- fetch/cache 系テストを ProgramFetcherTest (test/unit/lib) へ移設、
  ProgramTest には参照・編集・update 統合テストを残す
